Add view of org page

+++ public/js/app.js
@@ -1980,6 +1980,9 @@ __webpack_require__.r(__webpack_exports__);
 //
 //
 //
+//
+//
+//
 /* harmony default export */ __webpack_exports__["default"] = ({
   mounted: function mounted() {
     this.getOrgData();
@@ -1995,9 +1998,13 @@ __webpack_require__.r(__webpack_exports__);
     deleteOrg: function deleteOrg(id, index) {
       var _this2 = this;
 
-      axios.delete("api/listorgs/" + id).then(function (response) {
-        _this2.ListOrgs.splice(index, 1);
-      });
+      if (confirm("Do you really want to delete?")) {
+        axios.delete("api/listorgs/" + id).then(function (response) {
+          _this2.ListOrgs.splice(index, 1);
+        }).catch(function (error) {
+          console.log(error);
+        });
+      }
     }
   },
   data: function data() {
@@ -2161,6 +2168,62 @@ __webpack_require__.r(__webpack_exports__);
         rating: 0
       }
     };
+  }
+});
+
+/***/ }),
+
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/ListOrg/ShowOrg.vue?vue&type=script&lang=js&":
+/*!**************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/ListOrg/ShowOrg.vue?vue&type=script&lang=js& ***!
+  \**************************************************************************************************************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+/* harmony default export */ __webpack_exports__["default"] = ({
+  props: ["id"],
+  data: function data() {
+    return {
+      name_org: "",
+      picture: "",
+      headerpic: "",
+      description: ""
+    };
+  },
+  mounted: function mounted() {
+    var _this = this;
+
+    axios.get("/api/listorgs/" + this.id).then(function (response) {
+      var listorg = response.data;
+      _this.name_org = listorg.name_org;
+      _this.picture = listorg.picture;
+      _this.description = listorg.description;
+      _this.headerpic = listorg.headerpic;
+    });
+  },
+  methods: {
+    updateUser: function updateUser() {
+      axios.put("/api/listorgs/" + this.id, {
+        name_org: this.name_org
+      });
+    }
   }
 });
 
@@ -37534,9 +37597,14 @@ var render = function() {
                             [_vm._v("Delete")]
                           ),
                           _vm._v(" "),
-                          _c("h2", { staticClass: "txt-bold" }, [
-                            _vm._v(_vm._s(ListOrg.name_org))
-                          ]),
+                          _c(
+                            "a",
+                            {
+                              staticClass: "txt-bold text-link",
+                              attrs: { href: "/listorgs/" + ListOrg.id }
+                            },
+                            [_vm._v(_vm._s(ListOrg.name_org))]
+                          ),
                           _vm._v(" "),
                           _c(
                             "a",
@@ -37591,7 +37659,7 @@ var render = function() {
                     _c(
                       "label",
                       { staticClass: "title", attrs: { for: "sel1" } },
-                      [_vm._v("สายงานที่สนใจ : " + _vm._s(_vm.type) + " ")]
+                      [_vm._v("สายงานที่สนใจ : " + _vm._s(_vm.type))]
                     ),
                     _vm._v(" "),
                     _c(
@@ -37862,6 +37930,53 @@ var staticRenderFns = [
     ])
   }
 ]
+render._withStripped = true
+
+
+
+/***/ }),
+
+/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/ListOrg/ShowOrg.vue?vue&type=template&id=681cf443&":
+/*!******************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/ListOrg/ShowOrg.vue?vue&type=template&id=681cf443& ***!
+  \******************************************************************************************************************************************************************************************************************/
+/*! exports provided: render, staticRenderFns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "render", function() { return render; });
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return staticRenderFns; });
+var render = function() {
+  var _vm = this
+  var _h = _vm.$createElement
+  var _c = _vm._self._c || _h
+  return _c(
+    "div",
+    { staticClass: "container col-md-10 col-sm-9 col-lg-10 mb-5" },
+    [
+      _c("div", { staticClass: "card border-0 shadow" }, [
+        _c("img", {
+          staticStyle: {
+            height: "100%",
+            width: "100%",
+            "object-fit": "contain"
+          },
+          attrs: { src: "../images/data/List/" + _vm.headerpic }
+        }),
+        _vm._v(" "),
+        _c("div", { staticClass: "container mt-4 mb-3" }, [
+          _c("h2", { staticClass: "title" }, [_vm._v(_vm._s(_vm.name_org))]),
+          _vm._v(" "),
+          _c("h2", { staticClass: "txt-regular" }, [
+            _vm._v(_vm._s(_vm.description))
+          ])
+        ])
+      ])
+    ]
+  )
+}
+var staticRenderFns = []
 render._withStripped = true
 
 
@@ -50080,6 +50195,7 @@ Vue.component('orgs', __webpack_require__(/*! ./components/ListOrg/Orgs.vue */ "
 Vue.component('edit', __webpack_require__(/*! ./components/ListOrg/Edit.vue */ "./resources/js/components/ListOrg/Edit.vue").default);
 Vue.component('create', __webpack_require__(/*! ./components/ListOrg/Create.vue */ "./resources/js/components/ListOrg/Create.vue").default);
 Vue.component('search', __webpack_require__(/*! ./components/ListOrg/Search.vue */ "./resources/js/components/ListOrg/Search.vue").default);
+Vue.component('show-org', __webpack_require__(/*! ./components/ListOrg/ShowOrg.vue */ "./resources/js/components/ListOrg/ShowOrg.vue").default);
 var app = new Vue({
   el: '#app'
 });
@@ -50553,6 +50669,75 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_Search_vue_vue_type_template_id_4af75588___WEBPACK_IMPORTED_MODULE_0__["render"]; });
 
 /* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_Search_vue_vue_type_template_id_4af75588___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
+
+
+
+/***/ }),
+
+/***/ "./resources/js/components/ListOrg/ShowOrg.vue":
+/*!*****************************************************!*\
+  !*** ./resources/js/components/ListOrg/ShowOrg.vue ***!
+  \*****************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _ShowOrg_vue_vue_type_template_id_681cf443___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./ShowOrg.vue?vue&type=template&id=681cf443& */ "./resources/js/components/ListOrg/ShowOrg.vue?vue&type=template&id=681cf443&");
+/* harmony import */ var _ShowOrg_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./ShowOrg.vue?vue&type=script&lang=js& */ "./resources/js/components/ListOrg/ShowOrg.vue?vue&type=script&lang=js&");
+/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
+
+
+
+
+
+/* normalize component */
+
+var component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__["default"])(
+  _ShowOrg_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
+  _ShowOrg_vue_vue_type_template_id_681cf443___WEBPACK_IMPORTED_MODULE_0__["render"],
+  _ShowOrg_vue_vue_type_template_id_681cf443___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
+  false,
+  null,
+  null,
+  null
+  
+)
+
+/* hot reload */
+if (false) { var api; }
+component.options.__file = "resources/js/components/ListOrg/ShowOrg.vue"
+/* harmony default export */ __webpack_exports__["default"] = (component.exports);
+
+/***/ }),
+
+/***/ "./resources/js/components/ListOrg/ShowOrg.vue?vue&type=script&lang=js&":
+/*!******************************************************************************!*\
+  !*** ./resources/js/components/ListOrg/ShowOrg.vue?vue&type=script&lang=js& ***!
+  \******************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_ShowOrg_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/babel-loader/lib??ref--4-0!../../../../node_modules/vue-loader/lib??vue-loader-options!./ShowOrg.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/ListOrg/ShowOrg.vue?vue&type=script&lang=js&");
+/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_ShowOrg_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
+
+/***/ }),
+
+/***/ "./resources/js/components/ListOrg/ShowOrg.vue?vue&type=template&id=681cf443&":
+/*!************************************************************************************!*\
+  !*** ./resources/js/components/ListOrg/ShowOrg.vue?vue&type=template&id=681cf443& ***!
+  \************************************************************************************/
+/*! exports provided: render, staticRenderFns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_ShowOrg_vue_vue_type_template_id_681cf443___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../node_modules/vue-loader/lib??vue-loader-options!./ShowOrg.vue?vue&type=template&id=681cf443& */ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/ListOrg/ShowOrg.vue?vue&type=template&id=681cf443&");
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_ShowOrg_vue_vue_type_template_id_681cf443___WEBPACK_IMPORTED_MODULE_0__["render"]; });
+
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_ShowOrg_vue_vue_type_template_id_681cf443___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
